Merge Feature/Slurs (#359)
Squashed commits:

* Added slur reader

* Coded first version of calculateSlurs() in VexFlowMusicSheetCalculator. Added class VexFlowSlur in the equally named ts-file.
TODO: write correct call procedure for calculateSlurs(). Formally it was in MusicSheetCalculator but now it immediately generates VexFlow slurs and saves them to the VexFlow staffline.

* Added class definition of Vex.Flow.Curve in vexflow.d.ts
Added createVexFlowCurve(...) in VexFlowSlur.ts
Construct curve in check for Endnote of slur in calculateSlurs() in VexFlowMusicSheetCalculator.ts
Added drawSlurs(...:VexFlowStaffline) in VexFlowMusicSheetDrawer.ts
Closed Slurs work fine. Open slurs still throw an error (like in "An die ferne Geliebte" - Beethoven)

* Problem: in vexflow.d.ts in the definition of the class Stave (derived from Modifier), not only the interface of the function "public static get Position()" is given but also the return values defined.
Apparently in vexflow.d.ts only definitions without bodies are allowed.
I commented the function in vexflow.d.ts and added in VexFlowMeasure.ts the Enum StavePositionEnum that now handles the numbers and keywords. This makes sense, as the function was only used in this file.

* Corrected the position for removing VexFlowSlur from array vfSlurs

* First Version working by design:
store all slurs to be drawed to the VexFlowStaffline.
use vfOpenSlurs to store slurs that start/end in different stafflines
Moved createVexFlowCurve() to drawSlurs in VexFlowMusicSheetDrawer.ts in oder not to draw curves with undefined endpoints and run into an error.

* added translated ts-file GraphicalCurve.ts for generating points of a curve

* work in progress: files Matrix2D (rotation matrix), Graphical Slur and Sorter for Graphical Slur, translated from c# and syntax corrected (except Graphical Slur)

* Reworked GraphicalSlur.ts:
Old positioning routines have been removed or commented as they will be replaced with calculations using the bounding box and the skyline. Therefore "PositionAndShape" has been used for the moment for calculation (the "else" case of the old if statements)

* Added Comments from .cs files and structured code by newlines. In GraphicalSlur.ts there is also commented old redundant code.

* Implemented CalculatingSlurs function in VexFlowMusicSheetCalculator.ts using GraphicalSlurs. Code runs and seems correct in Debugger. At the moment no rendering of the Slurs is available, so they do not appear on the sheet in the browser.

* Added method for rendering bezier curve on context.
Fixed methods with ref parameters.
Fixed typo.
Added comments.

* Removed unecessary <number> casts.

* Slurs are drawn in a first version. But: Grace notes are wrong again. Curves are not correct.

* fixed curve rendering for slurs.

* Excluded drawing of crossed slurs (from one staffline to another) as the calculation is still missing for them.
Fixed placement issues for two voices.

* Fixed import bug