update generate:current and blessed npm commands, pin vexflow to a version, updage generateDiffImages
without pinning vexflow to a version, you can't set it to an older version and install again.
we want to pin vexflow to a version anyways, because it's critical for rendering differences.
generateDiffImages now goes over all samples except Actor prelude, and uses file name instead of sample name with spaces