feat(generatePNGs): can generate PNGs browserless by node script, improved speed (#670)
close #670

browserless PNG generation is about twice as fast as puppeteer,
and has a way faster startup (about 4x as fast for single sample)

see the new/changed npm scripts "generatePNG*" and "generate:current".

puppeteer is not used for main scripts anymore, because it is slower,
so puppeteer was removed as devDependency, as it needs to download >100MB for Chromium.
install it as dev dependency locally if you want to test generatePNG:puppeteer*.