fix(Cursor): position correctly on page 2+ (#817)
fix #817

add the height of previous pages to the y position of the cursor
this is unfortunately necessary right now, because MusicPages all have RelativePosition and AbsolutePosition y=0.
without this, the cursor wouldn't be y-positioned correctly at page numbers > 1.

the div elements of both SVG and Canvas backends now have ids "osmdCanvasPage{pageNumber}".