fix tests, Haydn multi rest error bar 128
#861

testing if a source measure for which the xml gave a multi rest should be reduced to a multi rest
is not a good idea at the moment, because the first multi rest measure will still be created.
also, this works fine for the Haydn.

the previous commit caused an infinite loop in the Haydn sample.