feat(Measure Numbers): Display measure numbers (labels) as given in XML (#541)
polish of PR #879

renders the measure numbers as given in the XML, which can be completely arbitrary.
Doesn't affect SourceMeasure.MeasureNumber, which is now the logical nth measure printed in the sheet.

This will cause `SourceMeasure.getPrintedMeasureNumber()` to return what was given in XML,
while `SourceMeasure.MeasureNumber` is unchanged and always starting at 1.
It makes sense to have one variable for measure numbers starting at 1 (or 0) representing the logical measure number, that is, "this measure is the n-th printed on screen", for array indexes etc. So the first measure printed on screen still always has `MeasureNumber` 1, while `MeasureNumberXML` and `MeasureNumberPrinted` (set in `getPrintedMeasureNumber()`) can be different.