fix(Container height): small scores don't significantly exceed bounding box anymore (SVG height). fix PageBottomMargin (#875) etc
remove EngravingRules.PageBottomExtraWhiteSpace, as PageBottomMargin was fixed and replaces it
fix #875 as the option isn't necessary anymore
improve solution for #788

PageBottomMargin can now be used to extend the white space below the score and extend the container (SVG) height.

fix for the SVG height is in OpenSheetMusicDisplay.createOrRefreshRenderBackend(),
where the height of the backend is now calculated from the bounding box plus margins (* 10 * zoom).

DrawingParameters default now resets EngravingRules (necessary to disable compacttight mode)

the SVGs generated by OSMD now have ids in the DOM (starting with osmdCanvasPage1)

EngravingRules.TitleTopDistance reduced to 5.0

add sample OSMD Function Test Container height, which is a small score
that had an element height of 300+ instead of the needed ~230
because of a too generous magic number.