fix(Layout): don't add first system's border to margin below title. Saves a lot of space.
the first system's border was added to the margin above it,
duplicating the role of the EngravingRules TitleBottomDistance (if title drawn)
and PageTopMargin.

this change saves a lot of space for the title part (header).

the old behavior can be mimicked by increasing PageTopMargin and/or TitleBottomDistance EngravingRules.

somewhat related to #898