fix(Layout): Pickup measures aren't unnecessarily wide anymore (#938)
fix #938

The length of pickup measures is now in proportion to the length of the pickup measure
compared to the active time signature.
E.g. a 1/4 pickup in a 3/4 time signature is now 1/4 / 3/4 = 1/3rd as wide.

This involves a temporary change regarding maxStaffEntries in VexFlowMusicSheetCalculator,
which were previously only considered for the top vertical measure (e.g. voice in voice+piano).
This should be considered for all measures (remove the if),
but that would change spacing overall,
and this commit represents the pure diff for pickup measures only.

Added two new test samples for pickup (implicit) measures