fix(Articulations): Don't assume placement above by default (#921) (e.g. staccato)
fix #921

this makes it so that when the placement is not given in the xml
(or the engravingrule that the xml setting should be used is disabled),
an articulation (e.g. staccato) is put below the note if the stem goes up.

add two test samples: test_articulation_staccato_placement(below|above)*