fix(Rests): Fix rest collisions with notes (y coordinate) (#621, #1076). Add EngravingRules.RestCollisionYPadding
see #1076 for a description of code changes.

add GraphicalNote.lineShift (shifts the line/y position of a rest note, or theoretically also any note i guess)
add EngravingRules.RestCollisionYPadding

add some test samples for rest collisions, since there were almost no changes in our existing samples

Basically, we check for the height of the lowest or highest note, depending on the rest positioning (voice 1 = above other notes, other voices = below), and put the rest further away a bit. Also, if there's a stem in the direction of the rest, we push the rest away further.

The solution is not ideal yet, because very unfortunately, at the time of the creation of the rests, we still don't have the simultaneous notes' correct bounding boxes, the beams, stem lengths, etc. - so basically we have to reconstruct/predict all of that.

* fix(Rests): Fix Y collisions (WIP, not regression tested yet)

* rest y positions: add some test samples

* fix render errors introduced in (2nd-)previous commit

* fix pitch undefined error (rest positioning)

* rest positioning: fix wrong note compared leading to skipping of positioning

* fix highest instead of lowest note compared for rest under upwards stem and vice versa

* quarter rests: slightly increase padding under up-stems (non-8th rests)

* fix positioning. rest collision avoidance should be functional now (at least in a few samples)

still needs regression testing

* remove console.log

* fix whole notes shifted too much (see function test voice alignment)

see OSMD Function Test - Voice Alignment

* increase rest shift for beamed notes, comment

* fix execution order

* remove branch samples

new samples are already on develop and merged here

* remove branch samples

* test: adjust xml work titles (title labels)

* comment

* comment

Co-authored-by: sschmidTU <s.schmid@phonicscore.com>