fix(Tuplets): Fix rest notes not having TypeLength set. Fix tuplet count not reset for next voice (rare issue) (#1207)
the tuplet count reset only matters when the next voice starts with the same tuplets.
Rest notes having TypeLength was required by the audio player,
where Instrument.Voices also has rest notes in the voice entries,
which in this base osmd version without audio player is not the case for some reason.

This detail also means that the tuplet number count doesn't increase in tuplets consisting of rests only here,
because the voices don't have these notes as voice entries.
This issue disappears in the audio player.

This change also fixes the whole rest at the start of Debussy - Mandoline.