fix(BoundingBox): Fix bounding box of MusicSystem and of SystemLeftLine too large (#1245)
SystemLeftLine bbox was too large, because it took BorderTop 0 (when the System BorderTop may be -1.3, thus "higher"),
and took the System's size (height) as BorderBottom, which is inaccurate either way.

Actually the bounding box is probably still too large, as the system left line shouldn't always be as large as the system,
but at least it's not larger now.