fix cursor update issue when making non-first instruments invisible after cursor set on now hidden position (#1312, #1310)
this was a silly assumption, starting from index 1, because not only the first instrument can be made invisible.
or rather, the default assignment to measure 0 was removed in a
refactor, making this start index invalid.

this caused an error when advancing to a position only visible in the
second instrument, then making that instrument invisible and
re-rendering.
e.g. Beethoven Geliebte advancing to the second note in the bass in
measure 2,
then osmd.Sheet.Instruments[1].Visible = false and osmd.render()