feat(MeasureWidth): Able to use same measure width for all measures (rules.FixedMeasureWidth). WIP. Also rules.FixedMeasureWidthFixedValue (experimental)
experimental.
the problem with fixed width is that to be 100% sure there are no layout issues, the maximum width of all measures needs to be used.
If Vexflow gets less space for a measure than it says it needs as a minimum, the layout can get completely messed up, collide, overflow, etc.
In a test, this already happened at around 66% of a measure's minimum width requested by Vexflow (just one data point)

To evaluate the measure widths more easily, use osmd.EngravingRules.RenderSingleHorizontalStaffline = true,
otherwise the measure widths might be changed to fit the systems (lines).