Fixed issue:  Chord symbols are associated with the wrong note #1270 of ErdoesiT. This issue also arised for the first time with a Dorico file.
I also refactored the 'InstrumentReader::readNextXmlMeasure()' method, just to understand it better and to declutter it a bit (I hope).

When the harmony tags are parsed, they are coupled to the correct staffEntry right away. Not in the next round. This seems to work with files that put the harmony tag before the note (chord test files that are included with osmd) as well as files that use backup, harmony, forward tags after a note (dorico files for instance).
Removed all dutch comments and commented console.logs