fix(Chord Symbols): Fix chord symbols associated with the wrong note/timestamp (#1270, PR #1342)
refactor: move else if for attributes node to the top (@johanvelthuis, see PR #1342)
stop using OpenChordSymbolContainers, which didn't handle backup nodes correctly -> wrong timestamp
refactor some xml handling into sub-methods (-> end of file/diff)

there will be one more commit after merging this to fix the missing space before the else.

thanks @johanvelthuis for the contribution, good changes!

merged commits:

* Fixed a problem where if slur and tie stops are in separate notations tags, the slur end would be missed, resulting in endless slurs on notes with both a slur end and a tie stop. According to the musicxml specs multiple notations tags are allowed, so they should be dealt with.

* Fixed issue:  Chord symbols are associated with the wrong note #1270 of ErdoesiT. This issue also arised for the first time with a Dorico file.

I also refactored the 'InstrumentReader::readNextXmlMeasure()' method, just to understand it better and to declutter it a bit (I hope).

When the harmony tags are parsed, they are coupled to the correct staffEntry right away. Not in the next round. This seems to work with files that put the harmony tag before the note (chord test files that are included with osmd) as well as files that use backup, harmony, forward tags after a note (dorico files for instance).

* Fixed issue:  Chord symbols are associated with the wrong note #1270 of ErdoesiT. This issue also arised for the first time with a Dorico file.

I also refactored the 'InstrumentReader::readNextXmlMeasure()' method, just to understand it better and to declutter it a bit (I hope).

When the harmony tags are parsed, they are coupled to the correct staffEntry right away. Not in the next round. This seems to work with files that put the harmony tag before the note (chord test files that are included with osmd) as well as files that use backup, harmony, forward tags after a note (dorico files for instance).
Removed all dutch comments and commented console.logs

* Removed a temporary testfile.

* Removed unnecessary spacings and line endings.

* Merged with develop branche

* Merged with develop branche