fix(X-Alignment): Fix note x-alignment for staves with different key signatures (PR #1427, #1425)
fixes #1425, #1315

This fix is an extended version of the one Vexflow did in their repository:
0xfe/vexflow#1390

In our case, this fix alone was not enough, because of how we use formatToStave(),
which overwrote the new x and start_x values after they had been set correctly.
(which was quite tricky to debug)

squashed commits:

* fix timesignature / note x-alignment for staves with different key signatures

it's unfortunate we may have to reference VexFlow classes in MusicSystemBuilder,
we're generally trying to avoid that.
Maybe we can rewrite that in the following commit with any etc

* refactor: remove Vexflow class references from MusicSystemBuilder

we generally don't want to reference Vexflow classes or our VexFlowX classes
(e.g. VexFlowMeasure) in base classes that are generally not specific to Vexflow.
E.g. referencing Vexflow things in VexFlowMeasure is fine, but should be avoided in GraphicalMeasure

* refactor: cleanup: remove unnecessary/commented code

* comment

* fix x-alignment of notes broken in second and further systems (because start_x is overwritten)

* comment

---------

Co-authored-by: sschmid <s.schmid@phonicscore.com>