feat(Notehead): Support inverted triangle notehead (PR #1428, #1418)
Note that the x and y positioning especially for the smaller notes isn't ideal, but hard to optimize. If we e.g. shift everything one more pixel, the triangle of the small notes is not well visible, basically "eaten up" by the stem, or not connected to it. There are some intricacies of the glyph code here because this is a custom glyph that wasn't originally supported by Vexflow, so without a real deep dive into the custom glyph code, I think this is as far as we can go right now. Probably the x-shift would need to scale with note scaling.

no visual regressions other than the new test samples.

squashed commits:

* feat(Notehead): Support inverted triangle notehead

* comments / VexFlowPatch doc

---------

Co-authored-by: sschmidTU <s.schmid@phonicscore.com>