feat(Lyrics): Make scores with lyrics shorter and more compact by improving long lyrics padding (#1394, PR #1395, PR #1474)
new EngravingRules:
LyricsUseXPaddingForLongLyrics
LyricsXPaddingReductionForLongNotes
LyricsXPaddingReductionForLastNoteInMeasure
LyricsXPaddingForLastNoteInMeasure

squashed commits:

* tune lyrics xPadding values for short notes with long text

* alternate padding calculation: additive instead of multiplicative. much tighter for 16th notes with long lyrics

* set XPaddingFactor = 1 instead of 0.8, rebalance around that

the factor 0.8 was making it so that long strings have too little padding, needing the awkward padding *= 1.15 workaround.
This also makes it less volatile to increase the factor e.g. to 1.1.

* add option to apply padding to last note in the measure as well (default false, usually unnecessary)

this is only necessary in rare cases where there are last notes in a measure with very long lyrics text

* comment

* lyrics spacing: enable for last note in measure by default, but reduce necessary padding for this case by 0.1

* lyrics spacing: don't add padding if next note has no lyrics

* simplify text width calculation, add ITextMeasurer.computeTextWidth()

this saves a multiplication per calculation, which is not much, but also simpler.

* revert simplification of text width calculation (shifted sheet title labels to the right for some reason)

* simplify measure width calculation (calculateXLayout), fix first measure calculated twice

* also apply lyrics padding to long notes, but reduced. increase it for short notes (lowers elongation factor)

the new threshold 1.7 instead of 2.2 reduces the elongation factor for Beethoven measure 2 from 1.42 to 1.14.

* refactor: remove `if` enclosure (was always true now)

this refactor was done in a separate commit instead of in the last one, because the diff for this is hard to read.
this just removes the if enclosure, nothing more.

* add debug code and info for wrong relativePosition.x issue. will be removed in next commit

* remove debug code from last commit

if we want to debug this again, we can always take the info/code from the last commit

* osmd.clear(): don't error when calling clear after rendering has failed

* fix half rests shifted in accidentals test by checking next staff entries instead of voice entries (voice.VoiceEntries has no rest notes)

fix for osmd function test accidentals (measure with half note with lyrics then half rest)

* fix BorderLeft added twice (probably no functional impact)

* assume we need padding at measure end

* comment console.log debug

* refactor: rename LyricsUseXPaddingForShortNotes to LyricsUseXPaddingForLongLyrics

since we also apply it to longer notes now (e.g. quarter)

* add EngravingRules LyricsXPaddingReductionForLongNotes and LyricsXPaddingReductionForLastNoteInMeasure

---------

Co-authored-by: sschmidTU <s.schmid@phonicscore.com>