feat(Pitch): Add ToStringShort method that returns a short string like "F#4" (F#, octave 4). Add getter, e.g. gNote.sourceNote.Pitch?.ToStringShortGet (#1510)
Pitch.ToString is very verbose, so it's nice to have a getter and method to return e.g. "A4" or "An5" for a note. (n = natural)
The getter can be useful for debugging, since you can just click on it in a debugging session.

Note that Pitch is undefined for rest notes, so take care to avoid null errors!